layout/base/tests/chrome/test_bug847890_paintFlashing.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for paint flashing</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script src="chrome://mochikit/content/tests/SimpleTest/WindowSnapshot.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/paint_listener.js"></script>

  <script type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    function startTest() {
      waitForAllPaintsFlushed(function () {
        var before = snapshotWindow(window, false);
        SpecialPowers.getDOMWindowUtils(window).paintFlashing = true;
        document.body.innerHTML = "bar";
        waitForAllPaintsFlushed(function () {
          document.body.innerHTML = "foo";
          waitForAllPaintsFlushed(function () {
            var after = snapshotWindow(window, false);
            ok(compareSnapshots(before, after, false)[0], "windows are different");
            SpecialPowers.getDOMWindowUtils(window).paintFlashing = false;
            SimpleTest.finish();
          });
        });
      });
    }
  </script>
</head>
<body onload="startTest()">foo</body>
</html>
